pyproject.toml
[tool.poetry]
name = "poetry-venv"
version = "1.0.0"
description = ""
authors = ["Mihail Babichev <54461515+mikle305@users.noreply.github.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
black = "^25.1.0"
django = "^5.2"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.9.1"
pydantic = "^2.11.5"


[tool.black]
line-length = 120


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
